Address PR feedback: refactor metric config architecture
- Move WorkoutConfigurationProvider to its own file
- Add onConfigurationChanged to HealthKitMetricConfigViewController
  that forwards from provider, so parent listens to config VC not provider
- Move workout type selection navigation logic into provider
- Add onConfigurationChanged to MetricConfigurationProvider protocol
