+++ BeeSwift/Settings/HealthKitMetricConfigViewController.swift
@@ -1,4 +1,3 @@
-import BeeKit
 import Foundation
 import SnapKit
 import UIKit
@@ -7,81 +6,10 @@ import UIKit
 
 protocol MetricConfigurationProvider: AnyObject {
   var numberOfRows: Int { get }
+  var onConfigurationChanged: (() -> Void)? { get set }
   func cell(for tableView: UITableView, at row: Int) -> UITableViewCell
   func didSelectRow(at row: Int)
   func getConfigParameters() -> [String: Any]
-}
-
-// MARK: - WorkoutConfigurationProvider
-
-class WorkoutConfigurationProvider: MetricConfigurationProvider {
-  let syncModeSegmentedControl = UISegmentedControl(items: ["Daily Total", "Individual Workouts"])
-
-  var selectedWorkoutTypes: [String] = []
-  var onConfigurationChanged: (() -> Void)?
-  var onNavigateToTypeSelection: (() -> Void)?
-
-  init(existingConfig: [String: Any] = [:]) {
-    if let types = existingConfig["workout_types"] as? [String] { selectedWorkoutTypes = types }
-    let initialDailyAggregate = existingConfig["daily_aggregate"] as? Bool ?? true
-    syncModeSegmentedControl.selectedSegmentIndex = initialDailyAggregate ? 0 : 1
-    syncModeSegmentedControl.addTarget(self, action: #selector(syncModeChanged), for: .valueChanged)
-  }
-
-  @objc private func syncModeChanged() { onConfigurationChanged?() }
-
-  func setSelectedWorkoutTypes(_ types: [String], in tableView: UITableView) {
-    selectedWorkoutTypes = types
-    tableView.reloadRows(at: [IndexPath(row: 1, section: 1)], with: .none)
-    onConfigurationChanged?()
-  }
-
-  // MARK: - MetricConfigurationProvider
-
-  var numberOfRows: Int { return 2 }
-
-  func cell(for tableView: UITableView, at row: Int) -> UITableViewCell {
-    if row == 0 {
-      // Sync Mode cell with segmented control
-      let cell = UITableViewCell(style: .default, reuseIdentifier: nil)
-      cell.selectionStyle = .none
-      cell.backgroundColor = .secondarySystemGroupedBackground
-
-      syncModeSegmentedControl.removeFromSuperview()
-      cell.contentView.addSubview(syncModeSegmentedControl)
-      syncModeSegmentedControl.snp.makeConstraints { make in
-        make.edges.equalToSuperview().inset(UIEdgeInsets(top: 8, left: 16, bottom: 8, right: 16))
-      }
-      return cell
-    } else {
-      // Workout Types cell with disclosure
-      let cell = UITableViewCell(style: .value1, reuseIdentifier: nil)
-      cell.textLabel?.text = "Workout Types"
-      cell.detailTextLabel?.text = workoutTypesDetailText()
-      cell.accessoryType = .disclosureIndicator
-      cell.backgroundColor = .secondarySystemGroupedBackground
-      return cell
-    }
-  }
-
-  func didSelectRow(at row: Int) { if row == 1 { onNavigateToTypeSelection?() } }
-
-  func getConfigParameters() -> [String: Any] {
-    let dailyAggregate = syncModeSegmentedControl.selectedSegmentIndex == 0
-    var params: [String: Any] = ["daily_aggregate": dailyAggregate]
-    if !selectedWorkoutTypes.isEmpty { params["workout_types"] = selectedWorkoutTypes }
-    return params
-  }
-
-  private func workoutTypesDetailText() -> String {
-    if selectedWorkoutTypes.isEmpty {
-      return "All Types"
-    } else if selectedWorkoutTypes.count == 1 {
-      return WorkoutActivityTypeInfo.find(byIdentifier: selectedWorkoutTypes[0])?.displayName ?? selectedWorkoutTypes[0]
-    } else {
-      return "\(selectedWorkoutTypes.count) types"
-    }
-  }
 }
 
 // MARK: - SelfSizingTableView
@@ -105,7 +33,11 @@ class HealthKitMetricConfigViewController: UIViewController {
   private let metricName: String
   var unitName: String? { didSet { tableView.reloadData() } }
 
-  var configurationProvider: MetricConfigurationProvider?
+  var onConfigurationChanged: (() -> Void)?
+
+  var configurationProvider: MetricConfigurationProvider? {
+    didSet { configurationProvider?.onConfigurationChanged = { [weak self] in self?.onConfigurationChanged?() } }
+  }
 
   init(goalName: String, metricName: String) {
     self.goalName = goalName
@@ -143,7 +75,7 @@ extension HealthKitMetricConfigViewController: UITableViewDelegate, UITableViewD
 
   func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
     if section == 0 {
-      return 3  // Goal Name, Apple Health Metric, Unit
+      return 3  // Goal, Metric, Unit
     } else {
       return configurationProvider?.numberOfRows ?? 0
     }
